+++ sync-for-payroll/CodatSyncPayroll/CodatSyncPayroll.cs
@@ -17,6 +17,8 @@ namespace Codat.Sync.Payroll
     using System.Threading.Tasks;
     using System;
 
+
+
     /// <summary>
     /// Sync for Payroll: The API for Sync for Payroll. <br/>
     /// 
@@ -88,6 +90,22 @@ namespace Codat.Sync.Payroll
     
     public class SDKConfig
     {
+        public static string[] ServerList = new string[]
+        {
+            "https://api.codat.io",
+        };
+        /// Contains the list of servers available to the SDK
+        public string serverUrl = "";
+        public int serverIndex = 0;
+
+        public string GetTemplatedServerDetails()
+        {
+            if (!String.IsNullOrEmpty(this.serverUrl))
+            {
+                return Utilities.TemplateUrl(Utilities.RemoveSuffix(this.serverUrl, "/"), new Dictionary<string, string>());
+            }
+            return Utilities.TemplateUrl(SDKConfig.ServerList[this.serverIndex], new Dictionary<string, string>());
+        }
     }
 
     /// <summary>
@@ -118,16 +136,12 @@ namespace Codat.Sync.Payroll
     public class CodatSyncPayroll: ICodatSyncPayroll
     {
         public SDKConfig Config { get; private set; }
-        public static List<string> ServerList = new List<string>()
-        {
-            "https://api.codat.io",
-        };
 
         private const string _language = "csharp";
-        private const string _sdkVersion = "3.0.0";
-        private const string _sdkGenVersion = "2.188.1";
+        private const string _sdkVersion = "3.0.1";
+        private const string _sdkGenVersion = "2.194.1";
         private const string _openapiDocVersion = "3.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 3.0.0 2.188.1 3.0.0 Codat.Sync.Payroll";
+        private const string _userAgent = "speakeasy-sdk/csharp 3.0.1 2.194.1 3.0.0 Codat.Sync.Payroll";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;
@@ -140,9 +154,14 @@ namespace Codat.Sync.Payroll
         public ICompanyInfo CompanyInfo { get; private set; }
         public ITrackingCategories TrackingCategories { get; private set; }
 
-        public CodatSyncPayroll(Security? security = null, string? serverUrl = null, ISpeakeasyHttpClient? client = null)
+        public CodatSyncPayroll(Security? security = null, int? serverIndex = null, string? serverUrl = null, Dictionary<string, string>? urlParams = null, ISpeakeasyHttpClient? client = null)
         {
-            _serverUrl = serverUrl ?? CodatSyncPayroll.ServerList[0];
+            if (serverUrl != null) {
+                if (urlParams != null) {
+                    serverUrl = Utilities.TemplateUrl(serverUrl, urlParams);
+                }
+                _serverUrl = serverUrl;
+            }
 
             _defaultClient = new SpeakeasyHttpClient(client);
             _securityClient = _defaultClient;
@@ -154,6 +173,7 @@ namespace Codat.Sync.Payroll
             
             Config = new SDKConfig()
             {
+                serverUrl = _serverUrl
             };
 
             Companies = new Companies(_defaultClient, _securityClient, _serverUrl, Config);