+++ sync-for-payroll/RELEASES.md
@@ -97,3 +97,13 @@ Based on:
 - [csharp v3.0.0] sync-for-payroll
 ### Releases
 - [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Sync.Payroll/3.0.0 - sync-for-payroll
+
+## 2023-11-20 01:30:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Payroll.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.1] sync-for-payroll
+### Releases
+- [NuGet v3.0.1] https://www.nuget.org/packages/Codat.Sync.Payroll/3.0.1 - sync-for-payroll